+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+- Network peering: add `WaitForNetworkPeeringState` helper
+
 ## [8.1.0]
 
 ### Added